refactor: move parse into the menuRouter
